Began work on building a King Kong based Pong
Created a new kong class that had the kong bouncing that external input
has no effect on.  Need to implement collision detection.  Need to
understand how to pass paddle position to kong class.

Started work on a paddle class that will use an empire state building
pic as the paddle.  Need to implement entire update function for this
class.  Will have two instances of this class for the two paddles.